Fixed tests, redefined API
