+++ src/sc2egset_dataset/dataset/lightning_datamodules/sc2_egset_datamodule.py
@@ -1,14 +1,11 @@
-from typing import Callable, List, Optional, Tuple
-
-import pytorch_lightning as pl
-from torch.utils.data import random_split
-from torch.utils.data.dataloader import DataLoader
-from sc2egset_dataset.dataset.available_replaypacks import AVAILABLE_REPLAYPACKS
-
-from sc2egset_dataset.dataset.pytorch_datasets.sc2_egset_dataset import SC2EGSetDataset
+from typing import Callable, List, Tuple
+from sc2egset_dataset.dataset.available_replaypacks import (
+    SC2EGSET_DATASET_REPLAYPACKS,
+)
+from sc2egset_dataset.dataset.lightning_datamodules.sc2_datamodule import SC2DataModule
 
 
-class SC2EGSetDataModule(pl.LightningDataModule):
+class SC2EGSetDataModule(SC2DataModule):
     """
     Defines a LightningDataModule abstraction for the
     SC2EGSet: StarCraft II Esport Game-State Dataset
@@ -43,6 +40,7 @@ class SC2EGSetDataModule(pl.LightningDataModule):
 
     def __init__(
         self,
+        replaypacks: List[Tuple[str, str]] = SC2EGSET_DATASET_REPLAYPACKS,
         download_dir: str = "./data/download",
         unpack_dir: str = "./data/unpack",
         download: bool = True,
@@ -52,71 +50,17 @@ class SC2EGSetDataModule(pl.LightningDataModule):
         num_workers: int = 0,
         unpack_n_workers: int = 16,
         validator: Callable | None = None,
-        replaypacks: List[Tuple[str, str]] = AVAILABLE_REPLAYPACKS,
     ):
-        super().__init__()
 
-        # PyTorch fields:
-        self.transform = transform
-        self.dims = dims
-        self.batch_size = batch_size
-        self.num_workers = num_workers
-
-        # Custom fields:
-        self.download_dir = download_dir
-        self.unpack_dir = unpack_dir
-        self.download = download
-        self.unpack_n_workers = unpack_n_workers
-        self.validator = validator
-
-        self.replaypacks = replaypacks
-
-    def prepare_data(self) -> None:
-        # download, split, etc...
-        # only called on 1 GPU/TPU in distributed
-        self.dataset = SC2EGSetDataset(
-            download=self.download,
-            download_dir=self.download_dir,
-            unpack_dir=self.unpack_dir,
-            transform=self.transform,
-            unpack_n_workers=self.unpack_n_workers,
-            names_urls=self.replaypacks,
+        super().__init__(
+            replaypacks=replaypacks,
+            download_dir=download_dir,
+            unpack_dir=unpack_dir,
+            download=download,
+            transform=transform,
+            dims=dims,
+            batch_size=batch_size,
+            num_workers=num_workers,
+            unpack_n_workers=unpack_n_workers,
+            validator=validator,
         )
-
-    def setup(self, stage: Optional[str] = None) -> None:
-        # make assignments here (val/train/test split)
-        # called on every process in DDP
-
-        total_length = len(self.dataset)
-        # Add these to be a parameter in the initialization:
-        # 16.(6)% of total entries will be used for testing:
-        test_length = int(total_length / 6)
-        # 10% of total entries will be used for validation
-        val_length = int(total_length / 10)
-        # everything else will be used for training
-        train_length = total_length - test_length - val_length
-
-        self.train_dataset, self.test_dataset, self.val_dataset = random_split(
-            self.dataset,
-            [train_length, test_length, val_length],
-        )
-
-    def train_dataloader(self) -> DataLoader:
-        return DataLoader(
-            self.train_dataset, batch_size=self.batch_size, num_workers=self.num_workers
-        )
-
-    def val_dataloader(self) -> DataLoader:
-        return DataLoader(
-            self.val_dataset, batch_size=self.batch_size, num_workers=self.num_workers
-        )
-
-    def test_dataloader(self) -> DataLoader:
-        return DataLoader(
-            self.test_dataset, batch_size=self.batch_size, num_workers=self.num_workers
-        )
-
-    def teardown(self, stage: Optional[str] = None) -> None:
-        # clean up after fit or test
-        # called on every process in DDP
-        return super().teardown(stage)
