+++ tests/test_cases/lightning_datamodules/sc2_datamodule_test.py
@@ -2,15 +2,16 @@ import os
 from pathlib import Path
 import shutil
 import unittest
+from sc2egset_dataset.dataset.lightning_datamodules.sc2_datamodule import SC2DataModule
 
-from sc2egset_dataset.dataset.pytorch_datasets.sc2_egset_dataset import SC2EGSetDataset
-from sc2egset_dataset.dataset.replay_data.sc2_replay_data import SC2ReplayData
 from sc2egset_dataset.dataset.utils.zip_utils import unpack_zipfile
 from tests.test_utils.test_utils import get_specific_asset, get_test_output_dir
+
 from tests.settings_test import TEST_REPLAYPACKS
 
 
-class SC2EGSetDatasetTest(unittest.TestCase):
+# REVIEW: Verify these tests
+class SC2EGSetDataModuleTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls) -> None:
         cls.test_replaypack_name = "2022_TestReplaypack"
@@ -46,43 +47,22 @@ class SC2EGSetDatasetTest(unittest.TestCase):
         if self.unpacked.exists():
             shutil.rmtree(path=self.unpacked.as_posix())
 
-    def test_parsing_test_files(self):
-
-        sc2egset_dataset = SC2EGSetDataset(
+    def test_unpack_datamodule(self):
+        sc2_egset_datamodule = SC2DataModule(
             unpack_dir=self.unpack_dir_path,
             download_dir=self.download_dir_path,
             download=False,
-            names_urls=[(self.test_replaypack_name, "")],
+            replaypacks=TEST_REPLAYPACKS,
         )
 
-        self.assertIsInstance(sc2egset_dataset, SC2EGSetDataset)
-        # Files were properly indexed:
-        self.assertNotEqual(len(sc2egset_dataset), 0)
+        self.assertIsInstance(sc2_egset_datamodule, SC2DataModule)
 
-        # Testing positive indexing:
-        sc2_replaydata_0 = sc2egset_dataset[0]
-        self.assertIsInstance(sc2_replaydata_0, SC2ReplayData)
-
-        # Testing Negative indexing:
-        sc2_replaydata_1 = sc2egset_dataset[-1]
-        self.assertIsInstance(sc2_replaydata_1, SC2ReplayData)
-
-        # There is only one replay loaded, so two indexing methods should return
-        # the same object:
-        self.assertEqual(hash(sc2_replaydata_0), hash(sc2_replaydata_1))
-
-    def test_downloading_replaypacks(self):
-
-        sc2egset_dataset = SC2EGSetDataset(
+    def test_download_unpack_datamodule(self):
+        sc2_egset_datamodule = SC2DataModule(
             unpack_dir=self.unpack_dir_path,
             download_dir=self.download_dir_path,
             download=True,
-            names_urls=TEST_REPLAYPACKS,
+            replaypacks=TEST_REPLAYPACKS,
         )
 
-        # Dataset was initialized:
-        self.assertIsInstance(sc2egset_dataset, SC2EGSetDataset)
-        # Files were properly indexed:
-        self.assertNotEqual(len(sc2egset_dataset), 0)
-        # It is possible to retrieve a single file by index:
-        self.assertIsInstance(sc2egset_dataset[0], SC2ReplayData)
+        self.assertIsInstance(sc2_egset_datamodule, SC2DataModule)
