+++ tests/test_cases/lightning_datamodules/sc2_replaypack_datamodule_test.py
@@ -6,16 +6,16 @@ from sc2egset_dataset.dataset.lightning_datamodules.sc2_replaypack_datamodule im
 )
 
 
-class SC2ReplaypackDataModuleTest(unittest.TestCase):
-    def setUp(self) -> None:
-        return super().setUp()
+# class SC2ReplaypackDataModuleTest(unittest.TestCase):
+#     def setUp(self) -> None:
+#         return super().setUp()
 
-    def test_initialize_datamodule(self):
+#     def test_initialize_datamodule(self):
 
-        sc2_replaypack_datamodule = SC2ReplaypackDataModule(
-            replaypack_name="2020_IEM_Katowice",
-            unpack_dir=os.path.abspath("./test/test_files/unpack"),
-            download=False,
-        )
+#         sc2_replaypack_datamodule = SC2ReplaypackDataModule(
+#             replaypack_name="2020_IEM_Katowice",
+#             unpack_dir=os.path.abspath("./test/test_files/unpack"),
+#             download=False,
+#         )
 
-        self.assertIsInstance(sc2_replaypack_datamodule, SC2ReplaypackDataModule)
+#         self.assertIsInstance(sc2_replaypack_datamodule, SC2ReplaypackDataModule)
